% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-inputs.R
\name{check_equal_length}
\alias{check_equal_length}
\title{Confirm that two vectors are equal length}
\usage{
check_equal_length(x, y, call = rlang::caller_env())
}
\arguments{
\item{x, y}{Numeric vectors to check for size equality}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
Checks that the two input vectors have the same length and prints a formatted
error if they do not
}
