% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-inputs.R
\name{check_NA}
\alias{check_NA}
\title{Confirm that vector has no NA values}
\usage{
check_NA(x, arg = rlang::caller_arg(x), call = rlang::caller_env())
}
\arguments{
\item{x}{Numeric vectors to check for NAs}

\item{arg}{An argument name as a string. This argument
will be mentioned in error messages as the input that is at the
origin of a problem.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
Checks that the input vector does not contain any NAs and prints a formatted
error if it does
}
