% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor-number.R
\name{factor_number}
\alias{factor_number}
\title{Convert numeric vector to factored variable}
\usage{
factor_number(x, levels, labels = levels)
}
\arguments{
\item{x}{Numeric vector to be converted to factored variable}

\item{levels}{Levels for the factored variable}

\item{labels}{Labels for the factored variable levels; defaults to the values
of the levels}
}
\value{
Factored vector created from \code{x} with input \code{levels} and \code{labels}
}
\description{
Convert a numeric vector to a factored variable with a specific set of
levels. Useful in cases where the possible levels for the vector are known
and it is necessary to remove small rounding or floating-point errors.
}
\details{
Function attempts to convert factor and character inputs to numeric first;
throws an error if that conversion introduces NAs. Uses the \code{\link[=round_list]{round_list()}}
function to first round each value to the closest number within the set of
possible levels, then maps each value to the matching factor level and label.
}
\examples{
x <- c(3.0003, 4.0002, 1.01, 2.02, 10.1, 5, 5.5, 7)
y <- c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10)
factor_number(x, levels = y, labels = y)
}
\seealso{
\code{\link[=round_list]{round_list()}} to round numeric values to the closest numbers from a
given list
}
