% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor-number.R
\name{fct2num}
\alias{fct2num}
\title{Convert factored variable to numeric}
\usage{
fct2num(x)
}
\arguments{
\item{x}{Factored vector to be converted}
}
\value{
A numeric vector converted from \code{x}
}
\description{
Wrapper function for the base R method of factor conversion to make the code
clearer in other functions
}
\examples{
x <- factor(c(1, 2, 2, 4, 3, 3), levels = c(1, 2, 3, 4))
fct2num(x)
}
