% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/null-arguments.R
\name{null-drop}
\alias{null-drop}
\alias{\%||\%}
\title{Defaults for NULL values}
\usage{
a \%||\% b
}
\arguments{
\item{a}{The default output}

\item{b}{The output if \code{a} is \code{NULL}}
}
\description{
Outputs \code{a} unless \code{a} is NULL, then outputs \code{b}. Helpful if an argument is
optional and there is a means to calculate the default
}
\examples{
input <- NULL
x <- input \%||\% sum(c(1,2,3))
}
