% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in-near.R
\name{pair_near}
\alias{pair_near}
\title{Compare values within a tolerance}
\usage{
pair_near(x, y, tol = 1e-12)
}
\arguments{
\item{x, y}{A pair of numeric vectors to be compared pairwise}

\item{tol}{Relative tolerance of comparison; default value is \code{1e-12}}
}
\value{
A logical vector the same length as \code{x} and \code{y} indicating
where the paired values are equal within the relative tolerance
}
\description{
Determines if elements in two numbers or vectors are (pairwise) equal within
a relative tolerance value.
}
\details{
This function is similar to \code{\link[dplyr]{near}} from
\code{\link[dplyr]{dplyr}} except that it uses relative tolerance instead of
absolute tolerance. Uses \code{1e-12} as the default tolerance because the
\code{ismembertol()} function from MATLAB uses \code{1e-12} as the tolerance
for double-precision floating point numbers.
}
\examples{
x <- c(1000, 1000, 100, 100, 10, 10, 1, 1)
y <- x + x * c(5e-12, 5e-16, 5e-12, 5e-16, 5e-12, 5e-16, 5e-12, 5e-16)
pair_near(x, y)
}
\seealso{
\code{\link[=in_near]{in_near()}} and \link{\%~\%} for determining if values are members of a
vector within a relative tolerance
}
