% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace-na.R
\name{replace_all_na}
\alias{replace_all_na}
\title{Replace NA values in data frame}
\usage{
replace_all_na(df, replace = 0)
}
\arguments{
\item{df}{Data frame with NA values to be replaced}

\item{replace}{Value to replace all NAs with}
}
\value{
Data frame with NA values replaced with the \code{replace} value
}
\description{
Replace all NA values in a data frame with a specific value (defaults to 0).
}
\details{
Applies \code{\link[tidyr:replace_na]{tidyr::replace_na()}} to all columns in the data frame to replace all
NAs with the given value.
}
\examples{
df <- data.frame(
  A = c(1, NA, 2, 3),
  B = c(NA, 2, NA, 3),
  C = c(1, 2, 3, 4),
  D = c(NA, NA, NA, 2)
)
replace_all_na(df, replace = 0)

}
\seealso{
\code{\link[tidyr:replace_na]{tidyr::replace_na()}} replaces all NA values in a specific column
with a given value.
}
