% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor-number.R
\name{round_list}
\alias{round_list}
\title{Round numbers to the closest values in a list}
\usage{
round_list(values, round_to)
}
\arguments{
\item{values}{Numeric vector with values that should be rounded}

\item{round_to}{Numeric vector with that the \code{values} should be rounded to}
}
\value{
A numeric vector with the \code{values} rounded to the closest number in
\code{round_to}
}
\description{
Rounds a numeric vector to the nearest values from a list of arbitrary
numbers. Helpful to remove small rounding or floating-point errors when the
list of possible values for the vector is known.
}
\details{
Used as the first step in converting a numeric vector to a factor in the
\code{\link[=factor_number]{factor_number()}} function. Originally adapted from a \href{https://stackoverflow.com/questions/12861061/}{StackOverflow post}.
}
\examples{
x <- c(3.0003, 4.0002, 1.01, 2.02, 10.1, 5, 5.5, 7)
y <- c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10)
round_list(values = x, round_to = y)
}
\seealso{
\code{\link[=factor_number]{factor_number()}} converts numeric vectors to factored vectors with
the numbers as the factor levels
}
