% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small-list.R
\name{small}
\alias{small}
\title{Smaller list preview}
\usage{
small(x, level = 4)
}
\arguments{
\item{x}{A list or list-type object (e.g., a data frame)}

\item{level}{The maximum number of list levels to display, defaults to 4}
}
\value{
The output of the \code{str()} function with specific options
}
\description{
Smaller list preview
}
\examples{
x <- list(a = 1:3, b = list(c = 5:9))
small(x)
}
